Renamed Param component to Node
