+++ src/containers/Macro/index.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux'
-import Param from '../../components/Param'
+import Node from '../../components/Node'
 import { sketchNodeOpenedToggle } from '../../store/sketches/actions'
 import getIsSketchNodeOpened from '../../selectors/getIsSketchNodeOpened'
 
@@ -18,6 +18,6 @@ const mapDispatchToProps = (dispatch, ownProps) => ({
 const ParamContainer = connect(
   mapStateToProps,
   mapDispatchToProps
-)(Param)
+)(Node)
 
 export default ParamContainer
